refactor(client): use shorts instead of bytes to avoid unnecessary conversions in audio capture processing

+++ api/client/src/main/java/su/plo/voice/api/client/audio/device/InputDevice.java
@@ -24,14 +24,14 @@ public interface InputDevice extends AudioDevice {
      * @param bufferSize the frame size
      * @return samples or null if bufferSize > available samples
      */
-    byte[] read(int bufferSize);
+    short[] read(int bufferSize);
 
     /**
      * Read samples with the capacity of bufferSize calculated from AudioFormat
      *
      * @return samples or null if bufferSize > available samples
      */
-    byte[] read();
+    short[] read();
 
     @Override
     default DeviceType getType() {
